formats vm-nogui: resize only if /dev/tty exists

+++ formats/vm-nogui.nix
@@ -1,12 +1,14 @@
 { pkgs, ... }:
 let
   resize = pkgs.writeScriptBin "resize" ''
-    old=$(stty -g)
-    stty raw -echo min 0 time 5
-    printf '\033[18t' > /dev/tty
-    IFS=';t' read -r _ rows cols _ < /dev/tty
-    stty "$old"
-    stty cols "$cols" rows "$rows"
+    if [ -e /dev/tty ]; then
+      old=$(stty -g)
+      stty raw -echo min 0 time 5
+      printf '\033[18t' > /dev/tty
+      IFS=';t' read -r _ rows cols _ < /dev/tty
+      stty "$old"
+      stty cols "$cols" rows "$rows"
+    ]
   ''; # https://unix.stackexchange.com/questions/16578/resizable-serial-console-window
 in {
   imports = [
